requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
openai>=1.0.0
qdrant-client>=1.11.0
fastembed>=0.4.0
redis>=5.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
